Forge: Moved memory and llm out of the sdk
