+++ autogpts/forge/forge/agent.py
@@ -7,13 +7,9 @@ from forge.sdk import (
     Task,
     TaskRequestBody,
     Workspace,
-    PromptEngine,
-    chat_completion_request,
-    ChromaMemStore,
 )
 from forge.actions import ActionRegister
-import json
-import pprint
+
 
 LOG = ForgeLogger(__name__)
