+++ autogpts/forge/forge/sdk/__init__.py
@@ -5,7 +5,7 @@ core of the Forge.
 from .agent import Agent
 from .db import AgentDB, Base
 from .forge_log import ForgeLogger
-from .llm import chat_completion_request, create_embedding_request, transcribe_audio
+from ..llm import chat_completion_request, create_embedding_request, transcribe_audio
 from .prompting import PromptEngine
 from .model import (
     Artifact,
@@ -23,5 +23,3 @@ from .model import (
 )
 from .workspace import LocalWorkspace, Workspace
 from .errors import *
-from .memory.chroma_memstore import ChromaMemStore
-from .memory.memstore import MemStore
